todoapp/requirement.txt
amqp==1.4.9
anyjson==0.3.3
apturl==0.5.2
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.3.0.23
bleach==3.0.2
Brlapi==0.6.6
celery==3.1.26.post2
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.7
django-celery==3.2.2
django-filter==2.1.0
djangorestframework==3.9.2
entrypoints==0.2.3
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
kombu==3.0.37
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.2
numexpr==2.6.4
numpy==1.15.4
oauth==1.0.1
olefile==0.45.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.0.0
psycopg2==2.7.7
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.0
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.13.2
six==1.11.0
soupsieve==1.8
statsmodels==0.9.0
sympy==1.3
system-service==0.3
systemd-python==234
tables==3.4.2
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xkit==0.0.0
zope.interface==4.3.2
